refactor(YALB-639): namespaced all atoms
