+++ components/01-atoms/atoms.scss
@@ -1,10 +1,10 @@
-@forward './controls/button/button' as button-*;
+@forward './controls/button/yds-button' as button-*;
 @forward './controls/cta/yds-cta';
 @forward './controls/text-link/yds-text-link';
-@forward './date-time/date-time';
-@forward './divider/divider';
-@forward './images/icons/icons';
-@forward './images/image/image';
-@forward './lists/list' as list-*;
-@forward './typography/headings/headings';
-@forward './typography/text/text';
+@forward './date-time/yds-date-time';
+@forward './divider/yds-divider';
+@forward './images/icons/yds-icons';
+@forward './images/image/yds-image';
+@forward './lists/yds-list' as list-*;
+@forward './typography/headings/yds-headings';
+@forward './typography/text/yds-text';
