+++ components/01-atoms/images/icons/yds-icons.twig
@@ -1,7 +1,7 @@
 <div class="icons-demo" id="icons">
   {% for item in icons %}
     <div class="icons-demo__item">
-      {% include "@atoms/images/icons/_icon.twig" with {
+      {% include "@atoms/images/icons/_yds-icon.twig" with {
         icon__name: item
       } %}
       <pre>{{ item }}</pre>
