+++ components/01-atoms/images/image/yds-image.twig
