+++ components/01-atoms/lists/yds-list.twig
@@ -15,7 +15,7 @@
 <{{ list__type }} {{ add_attributes(list__attributes) }}>
   {% block list__content %}
     {% for list__item in list__items %}
-      {% include "@atoms/lists/_list-item.twig" with {
+      {% include "@atoms/lists//_yds-list-item.twig" with {
         list__item__content: list__item.list__item__content,
         list__item__blockname: list__base_class,
       } %}
