+++ components/02-molecules/accordion/_accordion-item.twig
@@ -16,7 +16,7 @@
 
 {% set accordion__item__cta__content %}
   {{ accordion__item__heading }}
-  {% include "@atoms/images/icons/_icon.twig" with {
+  {% include "@atoms/images/icons/_yds-icon.twig" with {
     icon__name: 'angle-down',
     icon__blockname: accordion__item__base_class,
     icon__decorative: true,
@@ -33,7 +33,7 @@
 {% endset %}
 
 <div {{ add_attributes(accordion__item__attributes) }} >
-  {% include "@atoms/typography/headings/heading.twig" with {
+  {% include "@atoms/typography/headings/yds-heading.twig" with {
     heading__level: accordion__item__heading__level|default('2'),
     heading__blockname: accordion__item__base_class,
     heading: accordion__item__heading__content,
