+++ components/02-molecules/accordion/accordion.twig
@@ -10,7 +10,7 @@
 
 {% set expand_all__content -%}
   Expand All
-  {% include "@atoms/images/icons/_icon.twig" with {
+  {% include "@atoms/images/icons/_yds-icon.twig" with {
     icon__name: 'angle-down',
     icon__blockname: accordion__base_class,
   } %}
@@ -18,7 +18,7 @@
 
 {% set collapse_all__content -%}
   Collapse All
-  {% include "@atoms/images/icons/_icon.twig" with {
+  {% include "@atoms/images/icons/_yds-icon.twig" with {
     icon__name: 'angle-down',
     icon__blockname: accordion__base_class,
   } %}
@@ -32,19 +32,19 @@
 <div {{ add_attributes(accordion__attibutes) }} data-embedded-components>
   <div {{ bem('inner', [], accordion__base_class) }}>
     {% if accordion__heading %}
-      {% include "@atoms/typography/headings/heading.twig" with {
+      {% include "@atoms/typography/headings/yds-heading.twig" with {
         heading__level: '2',
         heading__blockname: accordion__base_class,
         heading: accordion__heading,
       } %}
     {% endif %}
-    {% embed "@atoms/lists/list.twig" with {
+    {% embed "@atoms/lists/yds-list.twig" with {
       list__base_class: 'controls',
       list__blockname: accordion__base_class,
       list__attributes: {'aria-label': 'Section controls'},
     } %}
       {% block list__content %}
-        {% embed "@atoms/lists/_list-item.twig" %}
+        {% embed "@atoms/lists//_yds-list-item.twig" %}
           {% block list__item__content %}
             {% include "@atoms/controls/base/yds-control.twig" with {
               control__content: expand_all__content,
@@ -54,7 +54,7 @@
             } %}
           {% endblock %}
         {% endembed %}
-        {% embed "@atoms/lists/_list-item.twig" %}
+        {% embed "@atoms/lists//_yds-list-item.twig" %}
           {% block list__item__content %}
             {% include "@atoms/controls/base/yds-control.twig" with {
               control__content: collapse_all__content,
