+++ components/02-molecules/banner/cta/cta-banner.twig
@@ -29,13 +29,13 @@
     </div>
     <div {{ bem('content', [], banner__base_class) }} data-component-theme={{ banner__content__background }}>
       <div {{ bem('text', [], banner__base_class) }}>
-        {% include "@atoms/typography/headings/heading.twig" with {
+        {% include "@atoms/typography/headings/yds-heading.twig" with {
           heading__level: '1',
           heading__blockname: banner__base_class,
           heading: banner__heading,
         } %}
         {% if banner__snippet %}
-          {% include "@atoms/typography/text/text.twig" with {
+          {% include "@atoms/typography/text/yds-text.twig" with {
             text__base_class: 'snippet',
             text__blockname: banner__base_class,
             text__content: banner__snippet,
