+++ components/02-molecules/cards/reference-card/examples/_card--examples.twig
@@ -1,12 +1,12 @@
 {% set reference_card__date__formatted %}
   {% set date_time__format = card_example_type == 'news' ? 'date' %}
-  {% include "@atoms/date-time/date-time.twig" with {
+  {% include "@atoms/date-time/yds-date-time.twig" with {
     date_time__start: reference_card__date,
   } %}
 {% endset %}
 
 {% set reference_card__time__formatted %}
-  {% include "@atoms/date-time/date-time.twig" with {
+  {% include "@atoms/date-time/yds-date-time.twig" with {
     date_time__start: reference_card__date,
     date_time__format: 'time',
   } %}
