+++ components/02-molecules/cards/reference-card/reference-card.twig
@@ -19,27 +19,27 @@
 <li {{ bem(reference_card__base_class) }}>
   <div {{ bem('content', [], reference_card__base_class) }}>
     {% if reference_card__overline %}
-      {% include "@atoms/typography/text/text.twig" with {
+      {% include "@atoms/typography/text/yds-text.twig" with {
         text__base_class: 'overline',
         text__blockname: reference_card__base_class,
         text__content: reference_card__overline,
       } %}
     {% endif %}
-    {% include "@atoms/typography/headings/heading.twig" with {
+    {% include "@atoms/typography/headings/yds-heading.twig" with {
       heading__level: reference_card__heading_level|default('3'),
       heading__blockname: reference_card__base_class,
       heading: reference_card__heading,
       heading__url: reference_card__url,
     } %}
     {% if reference_card__subheading %}
-      {% include "@atoms/typography/text/text.twig" with {
+      {% include "@atoms/typography/text/yds-text.twig" with {
         text__blockname: reference_card__base_class,
         text__base_class: 'subheading',
         text__content: reference_card__subheading,
       } %}
     {% endif %}
     {% if reference_card__snippet %}
-      {% include "@atoms/typography/text/text.twig" with {
+      {% include "@atoms/typography/text/yds-text.twig" with {
         text__base_class: 'snippet',
         text__blockname: reference_card__base_class,
         text__content: reference_card__snippet,
