+++ components/02-molecules/menu/_menu-list.twig
@@ -9,7 +9,7 @@
 {% endset %}
 
 {# List #}
-{% embed "@atoms/lists/list.twig" with {
+{% embed "@atoms/lists/yds-list.twig" with {
   list__base_class: 'menu',
   list__blockname: menu__base_class,
   list__modifiers: menu__modifiers,
