+++ components/02-molecules/meta/basic-meta/basic-meta.twig
@@ -1,4 +1,4 @@
-{% include "@atoms/typography/text/text.twig" with {
+{% include "@atoms/typography/text/yds-text.twig" with {
   text__content: basic_meta,
   text__base_class: 'basic-meta',
 } %}
