+++ components/02-molecules/meta/event-meta/event-meta.twig
@@ -7,7 +7,7 @@
  #}
 
 {% set event_meta__day %}
-  {% include "@atoms/date-time/date-time.twig" with {
+  {% include "@atoms/date-time/yds-date-time.twig" with {
     date_time__start: event_meta__date_start,
     date_time__end: event_meta__date_end,
     date_time__format: 'date',
@@ -15,7 +15,7 @@
 {% endset %}
 
 {% set event_meta__time %}
-  {% include "@atoms/date-time/date-time.twig" with {
+  {% include "@atoms/date-time/yds-date-time.twig" with {
     date_time__start: event_meta__date_start,
     date_time__end: event_meta__date_end,
     date_time__format: 'time',
