+++ components/02-molecules/page-title/page-title.stories.js
@@ -1,6 +1,6 @@
 // Twig templates
 import pageTitleTwig from './page-title.twig';
-import dateTimeTwig from '../../01-atoms/date-time/date-time.twig';
+import dateTimeTwig from '../../01-atoms/date-time/yds-date-time.twig';
 
 // Data files
 // import textData from './text/text.yml';
