+++ components/02-molecules/pager/pager.twig
@@ -8,7 +8,7 @@
         <li {{ bem('item', ['previous'], pager__base_class) }}>
           <a {{ bem('link', ['previous'], pager__base_class) }} href="{{ items.previous.href }}" title="{{ 'Go to previous page'|t }}" rel="prev"{{ items.previous.attributes|without('href', 'title', 'rel') }}>
             <span {{ bem('visually-hidden') }}>{{ 'Previous page'|t }}</span>
-            {% include "@atoms/images/icons/_icon.twig" with {
+            {% include "@atoms/images/icons/_yds-icon.twig" with {
               icon__name: 'angle-down',
               icon__decorative: true,
               icon__blockname: pager__base_class,
@@ -65,7 +65,7 @@
           <a {{ bem('link', ['next'], pager__base_class) }} href="{{ items.next.href }}" title="{{ 'Go to next page'|t }}" rel="next"{{ items.next.attributes|without('href', 'title', 'rel') }}>
             <span {{ bem('visually-hidden') }}>{{ 'Next page'|t }}</span>
             <span aria-hidden="true">{{ 'Next'|t }}</span>
-            {% include "@atoms/images/icons/_icon.twig" with {
+            {% include "@atoms/images/icons/_yds-icon.twig" with {
               icon__name: 'angle-down',
               icon__decorative: true,
               icon__blockname: pager__base_class,
