+++ components/02-molecules/social-links/_social-link__content.twig
@@ -2,7 +2,7 @@
   {# Visually hidden service name (e.g. Facebook) #}
   <span class="visually-hidden">{{ icon.name }}</span>
   {# Visually shown icon #}
-  {% include '@atoms/images/icons/_icon.twig' with {
+  {% include '@atoms/images/icons/_yds-icon.twig' with {
     icon__name: icon.icon,
     icon__decorative: true,
     icon__blockname: social_links__baseclass,
