+++ components/03-organisms/menu/breadcrumbs/breadcrumbs.twig
@@ -15,7 +15,7 @@
 <div {{ add_attributes(breadcrumbs__attributes) }}>
   {% set control__content %}
     <span {{ bem('visually-hidden') }}>Show all breadcrumbs</span>
-    {% include "@atoms/images/icons/_icon.twig" with {
+    {% include "@atoms/images/icons/_yds-icon.twig" with {
       icon__name: 'angle-down',
       icon__blockname: breadcrumbs__base_class,
       icon__decorative: true,
