+++ components/03-organisms/menu/utility-nav/utility-nav.twig
@@ -10,7 +10,7 @@
   {# Search #}
   {% if utility_nav__search %}
     <div {{ bem('search', [], utility_nav__base_class) }}>
-      {% include "@atoms/images/icons/_icon.twig" with {
+      {% include "@atoms/images/icons/_yds-icon.twig" with {
         icon__name: 'magnifying-glass',
         icon__blockname: utility_nav__base_class,
         icon__modifiers: ['search'],
