+++ components/05-page-examples/events/event-page.twig
@@ -22,7 +22,7 @@
     {% include "@molecules/text/text-field.twig" with {
       text_field__content: '<p>The political arena either on the local or national level is a territory where the presence of active women, despite constitutional guarantees, is considered an oddity. Especially when brave women with unveiled face don’t remain silent, but try to stimulate an open discussion in public space: a discussion on disproportions between the laws and the (in)compliance with them in terms of everyday life, on an open and latent discrimination of women, or on particular crimes of men against the women who try to live their life more dignified.</p><p>Part of a trilogy of documentaries by Afghan Director Sahraa Karimi screened from Tuesday 29th to Thursday 31st March, 2022.</p>',
     } %}
-    {% include "@atoms/divider/divider.twig" %}
+    {% include "@atoms/divider/yds-divider.twig" %}
     {% include "@molecules/text/text-field.twig" with {
       text_field__content: '
         <h2>Speaker</h2>
